WIP: Fix issue 135
Even though it doesn't change the interpolation result, the method
isTimepointStop of GtfsStopTime should be review. According to GTFS:
"Indicates if arrival and departure times for a stop are strictly
adhered to by the vehicle or if they are instead approximate and/or
interpolated times. This field allows a GTFS producer to provide
interpolated stop-times, while indicating that the times are
approximate. Valid options are:

0 - Times are considered approximate.
1 or empty - Times are considered exact."

That field is set to false to indicate it is
considered approximate when interpolate.